my app

+++ myapp/src/main/java/com/example/myapp/models/Product.java
@@ -3,9 +3,9 @@ package com.example.myapp.models;
 public class Product {
     private int id;
     private String name;
-    private float price;
+    private double price;
     
-    public Product(int id,String name,float price) {
+    public Product(int id,String name,double price) {
         this .id=id;
         this.name=name;
         this.price=price;
@@ -26,10 +26,10 @@ public class Product {
     public void setName(String name) {
         this.name = name;
     }
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
